Remove replaced code
